+++ en-SetSyncthingConfig.js
@@ -3,6 +3,7 @@
 
 // Notes:
 // * Generate config.xml if doesn't exist (firewall rule must already exist)
+// * If generating config for service, copy generated folder
 // * Set config.xml options requested by parameters
 // * Disable config.xml setLowPriority option
 
@@ -12,11 +13,14 @@ var MSG_FIREWALL_RULE_NOT_FOUND = "Firewall rule not found";
 var MSG_SYNCTHING_NOT_FOUND     = "syncthing.exe not found";
 var MSG_CONFIG_NOT_FOUND        = "File not found:";
 var MSG_CONFIG_NOT_UPDATED      = "Unable to update config.xml";
-var LOCALSERVICE_APPDATA_PATH   = "ServiceProfiles\\LocalService\\AppData\\Local";
 // END LOCALIZATION
+
+// LOCAL SERVICE root profile path inside SystemRoot
+var LOCALSERVICE_PROFILE_PATH = "ServiceProfiles\\LocalService";
 
 // Global Win32 API constants
 var ERROR_FILE_NOT_FOUND = 2;
+var ERROR_ALREADY_EXISTS = 183;
 var MB_ICONERROR         = 0x10;
 // Shell.Application constants
 var ssfLOCALAPPDATA = 28;
@@ -29,9 +33,10 @@ var ShellApp       = new ActiveXObject("Shell.Application");
 var WshShell       = new ActiveXObject("WScript.Shell");
 var XMLDOMDocument = new ActiveXObject("Microsoft.XMLDOM");
 // Global variables
-var ScriptPath             = WScript.ScriptFullName.substring(0,WScript.ScriptFullName.length - WScript.ScriptName.length);
-var CurrentUserConfigPath  = FSO.BuildPath(ShellApp.NameSpace(ssfLOCALAPPDATA).Self.Path,"Syncthing");
-var LocalServiceConfigPath = FSO.BuildPath(FSO.BuildPath(ShellApp.NameSpace(ssfWINDOWS).Self.Path,LOCALSERVICE_APPDATA_PATH),"Syncthing");
+var ScriptPath              = WScript.ScriptFullName.substring(0,WScript.ScriptFullName.length - WScript.ScriptName.length);
+var CurrentUserAppDataPath  = ShellApp.NameSpace(ssfLOCALAPPDATA).Self.Path;
+var LocalizedAppDataPath    = FSO.BuildPath(FSO.GetFileName(FSO.GetParentFolderName(CurrentUserAppDataPath)),FSO.GetFileName(CurrentUserAppDataPath));
+var LocalServiceAppDataPath = FSO.BuildPath(FSO.BuildPath(ShellApp.NameSpace(ssfWINDOWS).Self.Path,LOCALSERVICE_PROFILE_PATH),LocalizedAppDataPath);
 
 function help() {
   WScript.Echo("Usage: " + WScript.ScriptName + " {/currentuser|/localservice} <settings> [/silent]");
@@ -57,22 +62,50 @@ function paramIsEmpty(paramName) {
   return (Args.Named.Item(paramName) == "") || (Args.Named.Item(paramName) == null);
 }
 
+function copyFolder(source,destination) {
+  result = 0;
+  if ( destination.charAt(destination.length - 1) != "\\" ) {
+    destination += "\\";
+  }
+  try {
+    FSO.CopyFolder(source,destination,false);
+  }
+  catch(err) {
+    result = err.number;
+  }
+  return result;
+}
+
 function main() {
+  var currentUserConfigDir = FSO.BuildPath(CurrentUserAppDataPath,"Syncthing");
+  var currentUserConfigFileName = FSO.BuildPath(currentUserConfigDir,"config.xml");
+  var localServiceConfigDir = FSO.BuildPath(LocalServiceAppDataPath,"Syncthing");
+  var localServiceConfigFileName = FSO.BuildPath(localServiceConfigDir,"config.xml");
+
+  // Following depend on /currentuser or /localservice parameter
+  var configFileName = null;         // Path/filename of config.xml
+  var copyCurrentUserConfig = null;  // Copy current user config to service?
+  var defaultFolder = null;          // Add default folder to config?
+  var generate = null;               // Need to generate config.xml?
+
   if ( Args.Named.Exists("currentuser") ) {
-    var configDir = CurrentUserConfigPath;
-    var defaultFolder = true;
+    configFileName = currentUserConfigFileName;
+    copyCurrentUserConfig = false;
+    defaultFolder = true;
+    generate = ! FSO.FileExists(currentUserConfigFileName);
   }
   else if ( Args.Named.Exists("localservice") ) {
-    var configDir = LocalServiceConfigPath;
-    var defaultFolder = false;
+    configFileName = localServiceConfigFileName;
+    copyCurrentUserConfig = true;
+    defaultFolder = false;
+    generate = ! FSO.FileExists(localServiceConfigFileName);
   }
   else {
     help();
     return;
   }
 
-  var configFileName = FSO.BuildPath(configDir,"config.xml");
-  if ( ! FSO.FileExists(configFileName) ) {
+  if ( generate ) {
     var syncthingPath = FSO.BuildPath(ScriptPath,"syncthing.exe");
     // syncthing.exe chats on network, so abort if we can't find a firewall
     // rule for it (avoids Windows prompt to create firewall rules)
@@ -92,14 +125,25 @@ function main() {
     if ( ! defaultFolder ) {
       cmdLine += ' --no-default-folder';
     }
-    // Generate configuration
+    // Generate configuration (current user)
     WshShell.Run(cmdLine,0,true);
-    if ( ! FSO.FileExists(configFileName) ) {
+    if ( ! FSO.FileExists(currentUserConfigFileName) ) {
       if ( ! Args.Named.Exists("silent") ) {
-        WshShell.Popup(MSG_CONFIG_NOT_FOUND + "\n\n" + configFileName,0,MSG_DLG_TITLE,MB_ICONERROR);
+        WshShell.Popup(MSG_CONFIG_NOT_FOUND + "\n\n" + currentUserConfigFileName,0,MSG_DLG_TITLE,MB_ICONERROR);
       }
       return ERROR_FILE_NOT_FOUND;
     }
+    // Copy current user configuration to LOCAL SERVICE config
+    if ( copyCurrentUserConfig ) {
+      copyFolder(currentUserConfigDir,FSO.GetParentFolderName(localServiceConfigDir));
+    }
+  }
+
+  if ( ! FSO.FileExists(configFileName) ) {
+    if ( ! Args.Named.Exists("silent") ) {
+      WshShell.Popup(MSG_CONFIG_NOT_FOUND + "\n\n" + configFileName,0,MSG_DLG_TITLE,MB_ICONERROR);
+    }
+    return ERROR_FILE_NOT_FOUND;
   }
 
   // Set configuration options